pyproject.toml
[build-system]
requires = ["flit_core >=3.7.1,<4"]
build-backend = "flit_core.buildapi"

# if versions changed, change also on .pre-commit-config.yaml and .github/workflows/main.yaml
[project]
name = "firmatazero"
authors = [{name = "Olli Paloviita"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: BSD",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
readme = "README.md"
requires-python = ">=3.7"
dynamic = ["version", "description"]
keywords = ["firmatazero", "arduino", "gpiozero", "firmata", "pyFirmata2"]
dependencies = [
    "pyFirmata2==2.4.2",
    "pyserial==3.5",
]

[project.optional-dependencies]
dev = [
    "black==22.3.0",
    "codespell==2.1.0",
    "flake8==4.0.1",
    "isort==5.9.3",
    "pep8-naming==0.12.1",
    "pre-commit==2.18.1",
    "pyproject-flake8==0.0.1a2",
]

[project.urls]
Home = "https://github.com/ollipal/firmatazero"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79

[tool.black]
line-length = 79
exclude = '''
/(
    \.git
  | \.tox
  | \.venv
  | \.env
  | _build
  | build
  | docs/build
  | dist
  | .github
  | __pycache__
)/
'''

[tool.flake8]
ignore = "E203,E231,W503,I202"
max-line-length = 79
max-complexity = 18
select = "B,C,E,F,W,T4,B9,N"
exclude = ".git,.tox,.venv,.env,docs,dist,.github,__pycache__"
per-file-ignores = '''
    */__init__.py:F401,E501
'''